refactor(coding-agent): split ACP boundary helpers
