+++ packages/coding-agent/src/core/extensions/builtin/native-agent-sdk/acp-boundary.ts
@@ -1,16 +1,9 @@
 import { type ChildProcess, spawn } from "node:child_process";
 import type { Readable, Writable } from "node:stream";
-import {
-	client,
-	methods,
-	ndJsonStream,
-	PROTOCOL_VERSION,
-	type SessionUpdate,
-	type ToolCall,
-	type ToolCallUpdate,
-} from "@agentclientprotocol/sdk";
+import { client, methods, ndJsonStream, PROTOCOL_VERSION } from "@agentclientprotocol/sdk";
 import { reapProcessTree } from "../mcp/process-tree.ts";
-import { type NativeAgentPermissionRequest, requestNativeAgentPermission } from "./permission.ts";
+import { contentInput, type TrackedToolCall, trackToolCall } from "./acp-tool-calls.ts";
+import { requestNativeAgentPermission } from "./permission.ts";
 import type { NativeAgentEvent, NativeAgentRequest } from "./stream.ts";
 
 export type NativeAgentSessionConfigOption = {
@@ -90,44 +83,6 @@ class NativeAgentEventQueue implements AsyncIterableIterator<NativeAgentEvent> {
 		const reader = this.reader;
 		this.reader = undefined;
 		reader?.reject(error);
-	}
-}
-
-type TrackedToolCall = Pick<NativeAgentPermissionRequest, "kind" | "title" | "rawInput">;
-
-function contentInput(content: ToolCall["content"] | ToolCallUpdate["content"]): unknown {
-	const text = content
-		?.flatMap((item) => (item.type === "content" && item.content.type === "text" ? [item.content.text] : []))
-		.join("");
-	if (!text) return undefined;
-	try {
-		return JSON.parse(text);
-	} catch (error) {
-		if (error instanceof SyntaxError) return undefined;
-		throw error;
-	}
-}
-
-function trackToolCall(update: SessionUpdate, toolCalls: Map<string, TrackedToolCall>): void {
-	switch (update.sessionUpdate) {
-		case "tool_call":
-			toolCalls.set(update.toolCallId, {
-				kind: update.kind,
-				title: update.title,
-				rawInput: update.rawInput ?? contentInput(update.content),
-			});
-			return;
-		case "tool_call_update": {
-			const previous = toolCalls.get(update.toolCallId);
-			toolCalls.set(update.toolCallId, {
-				kind: update.kind ?? previous?.kind,
-				title: update.title ?? previous?.title ?? "Native agent tool request",
-				rawInput: update.rawInput ?? contentInput(update.content) ?? previous?.rawInput,
-			});
-			return;
-		}
-		default:
-			return;
 	}
 }
